created model module and updated functionality so that processed dates are tracked in a cache/file
